moved database commands out of main
